chore: move config struct to it's own file (#16779)
